+++ project-5/components/layout/header.tsx
@@ -1,7 +1,7 @@
 import useResponsive from '~hooks/use-responsive'
 
-import DesktopHeaderContent from './desktop-header-content'
-import MobileHeaderContent from './mobile-header-content'
+import DesktopHeaderContent from './desktop-header'
+import MobileHeaderContent from './mobile-header'
 
 export default function Header() {
   const isMobile = useResponsive('sm')
