+++ project-5/components/layout/index.tsx
@@ -1,0 +1,6 @@
+import Header from './header'
+import RoundedCorner from './rounded-corner'
+import ScrollWrapper from './scroll-wrapper'
+import withPageTransition from './with-page-transition'
+
+export { Header, RoundedCorner, ScrollWrapper, withPageTransition }
