+++ project-5/pages/projects/index.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 
 import { imageLoader } from 'common/utils'
-import withLayout from '~components/layout/with-layout'
+import { withPageTransition } from '~components/layout'
 
 const PROJECTS = ['project-1', 'project-2', 'project-3', 'project-4']
 
@@ -18,10 +18,7 @@ const Projects = () => {
 
       <div className="flex justify-center gap-8 my-auto">
         {PROJECTS.map((project, i) => (
-          <div
-            className="flex flex-col text-primary-zinc dark:text-primary-milk uppercase"
-            key={i}
-          >
+          <div className="flex flex-col text-primary-zinc dark:text-primary-milk uppercase" key={i}>
             <h2>{project}</h2>
 
             <div className="rounded overflow-hidden">
@@ -36,9 +33,7 @@ const Projects = () => {
               />
             </div>
 
-            <span className="text-sm text-right">
-              design style: scandinavian
-            </span>
+            <span className="text-sm text-right">design style: scandinavian</span>
             <span className="text-sm text-right">square meters: 78 sqm</span>
           </div>
         ))}
@@ -47,4 +42,4 @@ const Projects = () => {
   )
 }
 
-export default withLayout(Projects)
+export default withPageTransition(Projects)
